Refactored schedulers
