+++ .travis.yml
@@ -2,9 +2,8 @@ language: python
 sudo: false
 python:
 - 2.7
-- 3.3
-- 3.4
 - 3.5
+- 3.6
 - pypy
 before_install:
 - |
